fix: upcasting composite_subscription is unsafe
observable::subscribe returns a composite_subscription, which cannot be added to another composite_subscription, as ::add() only accepts a plain subscription. Upcasting a composite_subscription into a subscription loses its contents and would lead to leaks.

For now, we'll keep every subscription in a separate field, and explore using an array or set (or downclassing composite_subscription in a way that allows merging them?).